Refactor Firefox version conditional display
