+++ springfield/cms/templates/cms/blocks/icon-card-2026.html
@@ -11,7 +11,10 @@
 <include:conditional-display
   platform_conditions="{{ value.settings.show_to.platforms }}"
   firefox_condition="{{ value.settings.show_to.firefox }}"
-  auth_state_condition="{{ value.settings.show_to.auth_state }}">
+  auth_state_condition="{{ value.settings.show_to.auth_state }}"
+  min_version="{{ value.settings.show_to.min_version }}"
+  max_version="{{ value.settings.show_to.max_version }}"
+>
   <include:illustration-card-2026 variant="icon" expand_link="{{ value.settings.expand_link }}">
     <content:media>
       <include:icon
