+++ springfield/cms/templates/cms/blocks/illustration-card.html
@@ -10,7 +10,9 @@
 <include:conditional-display
   platform_conditions="{{ value.settings.show_to.platforms }}"
   firefox_condition="{{ value.settings.show_to.firefox }}"
-  auth_state_condition="{{ value.settings.show_to.auth_state }}">
+  auth_state_condition="{{ value.settings.show_to.auth_state }}"
+  min_version="{{ value.settings.show_to.min_version }}"
+  max_version="{{ value.settings.show_to.max_version }}">
   <include:illustration-card
     image_after="{{ value.settings.image_after }}"
     expand_link="{{ value.settings.expand_link }}"
