+++ springfield/cms/templates/cms/blocks/sections/home-carousel.html
@@ -20,7 +20,9 @@
 <include:conditional-display
   platform_conditions="{{ value.settings.show_to.platforms }}"
   firefox_condition="{{ value.settings.show_to.firefox }}"
-  auth_state_condition="{{ value.settings.show_to.auth_state }}">
+  auth_state_condition="{{ value.settings.show_to.auth_state }}"
+  min_version="{{ value.settings.show_to.min_version }}"
+  max_version="{{ value.settings.show_to.max_version }}">
   <include:carousel has_kit="true" items="{{ items }}">
     <content:heading>
       {% include_block value.heading %}
