+++ springfield/cms/templates/cms/blocks/sections/kit-banner.html
@@ -10,7 +10,9 @@
 <include:conditional-display
   platform_conditions="{{ value.settings.show_to.platforms }}"
   firefox_condition="{{ value.settings.show_to.firefox }}"
-  auth_state_condition="{{ value.settings.show_to.auth_state }}">
+  auth_state_condition="{{ value.settings.show_to.auth_state }}"
+  min_version="{{ value.settings.show_to.min_version }}"
+  max_version="{{ value.settings.show_to.max_version }}">
   <include:kit-banner theme="{{ value.settings.theme }}" anchor_id="{{ value.settings.anchor_id }}">
 
     <content:heading>
