+++ springfield/cms/templates/cms/blocks/sliding-carousel.html
@@ -9,7 +9,9 @@
 <include:conditional-display
   platform_conditions="{{ value.settings.show_to.platforms }}"
   firefox_condition="{{ value.settings.show_to.firefox }}"
-  auth_state_condition="{{ value.settings.show_to.auth_state }}">
+  auth_state_condition="{{ value.settings.show_to.auth_state }}"
+  min_version="{{ value.settings.show_to.min_version }}"
+  max_version="{{ value.settings.show_to.max_version }}">
   <section class="fl-section">
     <div class="fl-section-container">
       <div class="fl-sliding-carousel" data-js="fl-sliding-carousel">
