+++ springfield/cms/templates/cms/blocks/sticker-card-2026.html
@@ -10,7 +10,9 @@
 <include:conditional-display
   platform_conditions="{{ value.settings.show_to.platforms }}"
   firefox_condition="{{ value.settings.show_to.firefox }}"
-  auth_state_condition="{{ value.settings.show_to.auth_state }}">
+  auth_state_condition="{{ value.settings.show_to.auth_state }}"
+  min_version="{{ value.settings.show_to.min_version }}"
+  max_version="{{ value.settings.show_to.max_version }}">
   <include:sticker-card-2026 expand_link="{{ value.settings.expand_link }}">
     <content:pre_heading>{{ value.superheading|richtext|remove_p_tag }}</content:pre_heading>
