+++ springfield/cms/templates/pattern-library/components/conditional-display/conditional-display.yaml
@@ -1,16 +1,20 @@
-name: Conditional Display
+name: Conditional Display 2026
 context:
   variants:
     - name: All Users (no conditions)
       platform_conditions: []
       firefox_condition: null
       auth_state_condition: null
+      min_version: null
+      max_version: null
       contents: |
         <p>No condition classes are applied. Content is shown to all users.</p>
     - name: Firefox Users
       platform_conditions: []
       firefox_condition: is-firefox
       auth_state_condition: null
+      min_version: null
+      max_version: null
       html_class: is-firefox
       body_class:
       contents: |
@@ -20,6 +24,8 @@ context:
       platform_conditions: []
       firefox_condition: not-firefox
       auth_state_condition: null
+      min_version: null
+      max_version: null
       html_class: not-firefox
       body_class:
       contents: |
@@ -29,6 +35,8 @@ context:
       platform_conditions: []
       firefox_condition: null
       auth_state_condition: state-fxa-supported-signed-in
+      min_version: null
+      max_version: null
       html_class:
       body_class: state-fxa-supported-signed-in
       contents: |
@@ -38,6 +46,8 @@ context:
       platform_conditions: []
       firefox_condition: null
       auth_state_condition: state-fxa-supported-signed-out
+      min_version: null
+      max_version: null
       html_class:
       body_class: state-fxa-supported-signed-out
       contents: |
@@ -46,6 +56,10 @@ context:
     - name: Windows 10+ Users
       platform_conditions:
         - windows-10-plus
+      firefox_condition: null
+      auth_state_condition: null
+      min_version: null
+      max_version: null
       html_class: windows-10-plus
       body_class:
       contents: |
@@ -57,6 +71,8 @@ context:
         - linux
       firefox_condition: null
       auth_state_condition: null
+      min_version: null
+      max_version: null
       html_class: osx linux
       body_class:
       contents: |
@@ -69,19 +85,8 @@ context:
         - linux
       firefox_condition: is-firefox
       auth_state_condition: null
-      html_class: linux is-firefox
-      body_class:
-      contents: |
-        <p>Content is shown when <em>any</em> of the selected platform classes are present and the user is on Firefox.
-          Content with class <b>.conditional-display.condition-osx.condition-linux</b> wrapping content with class
-          <b>.conditional-display.condition-is-firefox</b> will be displayed if wrapped in class
-          <b>.osx</b> or <b>.linux</b> and <b>.is-firefox</b>.</p>
-    - name: macOS or Linux Users (multiple platforms) and Firefox Users
-      platform_conditions:
-        - osx
-        - linux
-      firefox_condition: is-firefox
-      auth_state_condition: null
+      min_version: null
+      max_version: null
       html_class: linux is-firefox
       body_class:
       contents: |
@@ -95,6 +100,8 @@ context:
         - ios
       firefox_condition: not-firefox
       auth_state_condition: null
+      min_version: null
+      max_version: null
       html_class: android not-firefox
       body_class:
       contents: |
@@ -102,3 +109,27 @@ context:
           Content with class <b>.conditional-display.condition-android.condition-ios</b> wrapping content with class
           <b>.conditional-display.condition-not-firefox</b> will be displayed if wrapped in class
           <b>.android</b> or <b>.ios</b> and <b>.not-firefox</b>.</p>
+    - name: Firefox >= 150 (minimum version)
+      platform_conditions: []
+      firefox_condition: null
+      auth_state_condition: null
+      min_version: 150
+      max_version: null
+      html_class:
+      body_class:
+      contents: |
+        <p>Content inside a <b>.conditional-display.condition-fx-version</b> div with
+          <b>data-min-version="150"</b> will be revealed by JS for Firefox >= 150
+          with the class <b>.version-match</b>.</p>
+    - name: Firefox <= 149 (maximum version)
+      platform_conditions: []
+      firefox_condition: null
+      auth_state_condition: null
+      min_version: null
+      max_version: 149
+      html_class:
+      body_class:
+      contents: |
+        <p>Content inside a <b>.conditional-display.condition-fx-version</b> div with
+          <b>data-max-version="149"</b> will be revealed by JS for Firefox <= 149
+          with the class <b>.version-match</b>.</p>
